backend/requirements.txt
langgraph
langchain-openai
langchain-core
pydantic
typing-extensions
openai
anthropic
python-dotenv
google-generativeai
